Added prompt for initial run of setup script.

+++ setup.sh
@@ -1,28 +1,32 @@
 #!/bin/bash
-sudo apt-get upgrade -y sudo apt-get update -y
 
-#==NODEJS==
-curl -sL https://deb.nodesource.com/setup_15.x | sudo -E bash -
-sudo apt-get install -y nodejs
+echo -n "Initial setup? (y/n)"
+read prompt
+if [ "$prompt" != "${prompt#[Yy]}" ] ;then
+  sudo apt-get upgrade -y sudo apt-get update -y
 
-
-#==MYSQL==
-sudo apt install mysql-server -y
-# TODO: Get this script to run without user input
-sudo mysql_secure_installation
-./create_schema.sh
-sudo mysql < create_schema.sh
+  #==NODEJS==
+  curl -sL https://deb.nodesource.com/setup_15.x | sudo -E bash -
+  sudo apt-get install -y nodejs
 
 
-#==NGINX==
-sudo apt-get install nginx -y
-sudo ufw allow 'Nginx HTTP'
-sudo ufw allow ssh
-sudo ufw allow 3000
-sudo ufw enable
-cd /etc/nginx/sites-enabled && sudo sed -i "s/root \/var\/www\/html;/root \/home\/ubuntu\/Flashcards\/frontend;/g" default
-sudo systemctl restart nginx
+  #==MYSQL==
+  sudo apt install mysql-server -y
+  # TODO: Get this script to run without user input
+  sudo mysql_secure_installation
+  ./create_schema.sh
+  sudo mysql < create_schema.sh
 
+
+  #==NGINX==
+  sudo apt-get install nginx -y
+  sudo ufw allow 'Nginx HTTP'
+  sudo ufw allow ssh
+  sudo ufw allow 3000
+  sudo ufw enable
+  cd /etc/nginx/sites-enabled && sudo sed -i "s/root \/var\/www\/html;/root \/home\/ubuntu\/Flashcards\/frontend;/g" default
+  sudo systemctl restart nginx
+fi
 
 #==HTTPS==
 echo -n "Set up HTTPS? (y/n)"
